+++ shared/src/main/scala/org/seekloud/thor/shared/ptcl/component/Food.scala
@@ -7,13 +7,13 @@ import org.seekloud.thor.shared.ptcl.model
 /**
   * Created by Jingyi on 2018/11/9
   */
-case class FoodState(fId: Long, level:Int, position: Point, radius: Float, color: Int)
+case class FoodState(fId: Long, level:Byte, position: Point, radius: Float, color: Int)
 
 trait Food extends CircleObjectOfGame {
 
   val fId: Long
 
-  var level: Int
+  var level: Byte
 
   var color: Int
 
@@ -34,7 +34,7 @@ object Food {
 
 case class AddNormalFood(
                         fId: Long,
-                        var level: Int,
+                        var level: Byte,
                         var position: model.Point,
                         var radius: Float,
                         var color: Int
